C3: refactor Next.js template code
Align the Next.js template code with the rest of the
C3 templates by removing its `templates.ts` file and
including templates files directly in the next template
directory
