+++ packages/create-cloudflare/templates/next/env.d.ts
@@ -1,0 +1,13 @@
+declare global {
+  namespace NodeJS {
+    interface ProcessEnv {
+      // Add here the Cloudflare Bindings you want to have available in your application
+      // (for more details on Bindings see: https://developers.cloudflare.com/pages/functions/bindings/)
+      //
+      // KV Example:
+      // MY_KV: KVNamespace
+    }
+  }
+}
+
+export {}
